+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 <a name="Pending Release"></a>
 ## [Pending Release](https://github.com/lightstep/otel-launcher-java/compare/0.17.1...master)
 * Remove the version number from agent jar file
-* Use OTEL_EXPORTER_OTLP_ENDPOINT environment variable instead of OTEL_EXPORTER_OTLP_SPAN_ENDPOINT
+* Use OTEL_EXPORTER_OTLP_TRACES_ENDPOINT environment variable instead of OTEL_EXPORTER_OTLP_SPAN_ENDPOINT
 
 <a name="0.17.1"></a>
 ## [0.17.1](https://github.com/lightstep/otel-launcher-java/compare/0.17.0...0.17.1)